inproc fork
